Add removal of word suggestions on long press.
There is a confirmation dialog, because all occurences of the suggestiong
will be removed. Removal affects only the (writable) user model(s), afterwards
the word may still exist in the system model(s).

The lm extension has been reworked to allow n-gram counts of 0. These n-grams
are considered removed and are ignored in predictions and when saving.
In memory, zeroed n-grams stay in n-gram trie and dictionary. Only saving and 
re-loading a model will finally purge them.